Move directory webui to portal as a sub-directory #58 (#61)
