+++ portal/webui/src/console/app.vue
