+++ portal/webui/src/console/components/keys.vue
