+++ portal/webui/src/console/components/networks.vue
