+++ portal/webui/src/console/main.js
