+++ portal/webui/src/login/components/auth_callback.vue
