+++ portal/webui/src/login/components/auth_redirect.vue
