--- webui/src/api/login.js
@@ -1,23 +1,0 @@
-// Copyright 2022 PairMesh, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-import service from "@/api/service";
-
-export function ssoMethods(query) {
-    return service.get('/api/v1/login/sso-methods' + query)
-}
-
-export function authCallback(provider, query) {
-    return service.post('/api/v1/login/auth/callback/'+provider + query)
-}
