--- webui/src/api/service.js
@@ -1,60 +1,0 @@
-// Copyright 2022 PairMesh, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-import axios from 'axios'
-
-const service = axios.create({
-    timeout: 5000,
-    responseType: 'json',
-    headers: {
-        "Content-Type": "application/json;charset=utf-8"
-    }
-})
-
-service.interceptors.request.use(
-    config => {
-        config.headers['Authorization'] = 'Bearer ' + localStorage.getItem('accessToken')
-        config.data = JSON.stringify(config.data);
-        return config;
-    },
-    error => {
-        return Promise.reject(error);
-    }
-)
-
-service.interceptors.response.use(
-    response => {
-        if (response.status === 200) {
-            return Promise.resolve(response);
-        } else {
-            return Promise.reject(response);
-        }
-    },
-    error => {
-        if (error.response.status) {
-            switch (error.response.status) {
-                case 401:
-                case 403:
-                    window.location.href = '/login';
-                    break;
-
-                default:
-                    return Promise.reject(error.response);
-            }
-            return Promise.reject(error.response);
-        }
-    },
-);
-
-export default service;
